refactor: changes architecture the agentruntime is used only library.
BREAKING CHANGE: Before the v0.2.1, the agentruntime is focus for
agentruntime to run by built-in binary file. However, it can be confused
for how developer uses agentruntime and network. So we pivot this
agentruntime as a library. Developer can use AgentRuntime as a
facade interfaces to build agents based on agentruntime library.